updated size_t as unsign long and add mem move

+++ ft_strlen.c
@@ -14,10 +14,10 @@
 
 size_t	ft_strlen(const char *s)
 {
-	unsigned int	i;
+	size_t	i;
 	
 	i = 0;
 	while (s[i] != '\0')
 		i++;
-	return ((size_t)i);
+	return (i);
 }
